requirements.txt
asgiref==3.8.1
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cssbeautifier==1.15.3
Django==5.1.6
django-environ==0.12.0
djangorestframework==3.15.2
djlint==1.36.4
EditorConfig==0.17.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
jsbeautifier==1.15.3
json5==0.10.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
smmap==5.0.2
sqlparse==0.5.3
tqdm==4.67.1
urllib3==2.3.0
